requirements.txt
# Core ML & Data Analysis
numpy==1.24.3
pandas==2.0.3
torch==2.0.1
xgboost==1.7.5
scikit-learn==1.3.0
optuna==3.2.0
shap==0.41.0
joblib==1.3.1
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.65.0

# Technical Analysis & Market Data
ta-lib==0.4.26
yfinance==0.2.18
twelvedata==1.2.14
alpha-vantage==2.3.1
requests==2.31.0

# API & Web Framework
flask==2.3.2
flask-cors==4.0.0
gunicorn==20.1.0
werkzeug==2.3.6

# Background Processing
celery==5.3.1
redis==4.6.0
kombu==5.3.1

# Database
sqlalchemy==2.0.19
psycopg2-binary==2.9.6
alembic==1.11.1

# Monitoring & Logging
prometheus-client==0.17.1
grafana-api==1.0.3

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Enterprise Feature Store Dependencies
clickhouse-connect==0.6.8
redis[hiredis]==4.6.0
msgpack==1.0.5
lz4==4.3.2
dask[complete]==2023.6.0
distributed==2023.6.0

# Advanced ML Libraries
evidently==0.4.9
ta==0.10.2
TA-Lib==0.4.26

# High-Performance Computing (Optional)
cudf-cu11==23.06.0; sys_platform != "darwin" and sys_platform != "win32"
cupy-cuda11x==12.1.0; sys_platform != "darwin" and sys_platform != "win32"
faiss-cpu==1.7.4

# Streaming & Messaging
kafka-python==2.0.2
apache-flink==1.17.1

# API Framework
fastapi==0.100.0
uvicorn==0.22.0
pydantic==2.0.0

# Additional Scientific Computing
scipy==1.11.1
statsmodels==0.14.0